Fix cross-compile in Dockerfile
The Dockerfile is using the `xx` tooling to select build targets.
This tooling is based on special Dockerfile arguments (TARGETOS,
TARGETARCH, etc).
These were missing from the image so we ended up building the wrong
target when we want a non-native platform.

Also removes the full `COPY . .` and replaces it with just copying the
data in that we need, which drastically improves caching.
In fact I found that the .dockerignore was not really working for us at
all and multiple back-to-back builds were having their caches
invalidated each time.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>